+++ src/components/base/Div/Div.tsx
@@ -5,17 +5,23 @@ export interface DivProps {
   children: React.ReactNode;
   margin?: string;
   padding?: string;
+  height?: string;
   width?: string;
   alignContent?: string;
+  alignItems?: string;
   justifyContent?: string;
 }
 
 export const StyledDiv = styled.div<DivProps>`
   ${(props) => props.margin && `margin: ${props.margin}`};
   ${(props) => props.padding && `padding: ${props.padding}`};
+  ${(props) => props.height && `height: ${props.height}`};
   ${(props) => props.width && `width: ${props.width}`};
-  ${(props) => (props.alignContent || props.justifyContent) && `display: flex`};
+  ${(props) =>
+    (props.alignContent || props.justifyContent || props.alignItems) &&
+    `display: flex`};
   ${(props) => props.alignContent && `align-content: ${props.alignContent}`};
+  ${(props) => props.alignItems && `align-items: ${props.alignItems}`};
   ${(props) =>
     props.justifyContent && `justify-content: ${props.justifyContent}`};
 `;
@@ -24,15 +30,19 @@ const Div = ({
   children,
   margin,
   padding,
+  height,
   width,
   alignContent,
+  alignItems,
   justifyContent,
 }: DivProps) => (
   <StyledDiv
     margin={margin}
     padding={padding}
+    height={height}
     width={width}
     alignContent={alignContent}
+    alignItems={alignItems}
     justifyContent={justifyContent}
   >
     {children}