+++ src/components/base/index.ts
@@ -4,3 +4,4 @@ export { default as Text } from "./Text";
 export { default as Paragraph } from "./Paragraph";
 export { default as Div } from "./Div";
 export { default as Img } from "./Img";
+export { default as IconTag } from "./IconTag";